Add test for handling file uploads
The view tests are worring, but the  serializer tests are not, presumably
because the encoding in a serializer is not set.
