+++ backend/backend/conftest.py
@@ -4,6 +4,8 @@ import factory
 import random
 from backend.users.models import User, Profile
 from backend.users.tests.factories import UserFactory, ProfileFactory, FakePhotoProfileFactory
+import shutil
+from pathlib import Path
 
 
 @pytest.fixture(autouse=True)
@@ -29,3 +31,13 @@ def profile_with_tags(user) -> Profile:
 @pytest.fixture
 def fake_photo_profile(user) -> Profile:
     return FakePhotoProfileFactory(user)
+
+
+
+@pytest.fixture
+def tmp_pic_path(tmp_path):
+    filename = "test_pic.png"
+    pic_path = Path().cwd() / 'backend' / 'users' / 'tests' / 'test_pic.png'
+    test_pic_path = tmp_path / filename
+    shutil.copy(pic_path, test_pic_path)
+    return test_pic_path
